tests: drivers: dma: scatter-gather PM runtime support enablement
Remove the log inside the DMA ISR to prevent assertion failures when
PM_DEVICE_RUNTIME is enabled and the console peripheral is managed by
PM runtime. This issue occurs due to a race condition.
Calling tc_print() from the ISR triggers pm_device_runtime_get() on the
peripheral (typically a UART device). Since these systems use
synchronization mechanisms, it can lead to taking a semaphore with
K_FOREVER while the system is in the PM state (PM_STATE_SUSPENDING).
An alternative would be to declare the PM device as ISR-safe using the
PM_DEVICE_FLAG_ISR_SAFE flag, but the PM device callback is not ISR-safe
in most UART drivers.

Signed-off-by: Martin Hoff <martin.hoff@silabs.com>